requirements.txt
torch
torchvision
numpy
opencv-python-headless
Pillow
PyYAML
tqdm
roboflow
matplotlib

